refactor: update prisma import paths to new structure
