+++ src/app/api/admin/event/[id]/route.test.ts
@@ -24,7 +24,7 @@ vi.mock('@/lib/helpers/prismaErrorHandler', () => ({
 
 // Import AFTER mocks so the route picks up the mocked deps
 import * as adminRoute from "@/app/api/admin/event/[id]/route";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/lib/prisma/prisma";
 import { adminCreateEventSchema, adminUpdateEventSchema } from "@/types/schemas/event/admin";
 import { handlePrismaError } from "@/lib/helpers/prismaErrorHandler";
