+++ src/app/api/auth/error/route.ts
@@ -1,7 +1,7 @@
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { cookies } from 'next/headers';
-import prisma from '../../../../lib/prisma';
+import prisma from '../../../../lib/prisma/prisma';
 import { createRedirectUrl } from '../../../../lib/auth/url';
 
 /**
