+++ src/app/api/event/[id]/route.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/lib/prisma/prisma";
 import { handlePrismaError } from "@/lib/helpers/prismaErrorHandler";
 import { type NextRequest, NextResponse } from "next/server";
 import { z } from "zod";
