+++ src/app/api/event/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/lib/prisma/prisma";
 import { handlePrismaError } from "@/lib/helpers/prismaErrorHandler";
