+++ src/app/api/user/route.ts
@@ -1,5 +1,5 @@
 import { type NextRequest, NextResponse } from "next/server";
-import prisma from "@/lib/prisma";
+import prisma from "@/lib/prisma/prisma";
 import { getSession } from "@/lib/auth/session";
 
 import { userSchema, createUserSchema } from "@/types/user";
