+++ src/app/auth/callback/link/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from 'next/navigation';
 import { getSession } from '@/lib/auth/session';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@/lib/prisma/prisma';
 
 interface LinkCallbackPageProps {
     searchParams: Promise<{
