--- src/lib/prisma.ts
@@ -1,9 +1,0 @@
-import { PrismaClient } from "@/generated/prisma";
-
-const globalForPrisma = globalThis as unknown as { prisma: PrismaClient };
-
-export const prisma = globalForPrisma.prisma || new PrismaClient();
-
-if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
-
-export default prisma;
