+++ src/lib/prisma/prisma.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from "@/generated/prisma";
+
+const globalForPrisma = globalThis as unknown as { prisma: PrismaClient };
+
+export const prisma = globalForPrisma.prisma || new PrismaClient();
+
+if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
+
+export default prisma;
