+++ src/lib/user/profilePicture.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/lib/prisma/prisma";
 import * as supa from "../supabase";
